Phase D: domain errors, services, full API surface
- Domain exceptions with JSON error_code/message handler
- Services: league create/reset/get, set user team, lineup validation,
  advance_matchday with single transaction commit
- API: /api/league, /api/teams, /api/teams/:id/{roster,lineup},
  /api/schedule, /api/standings, /api/games/:id, /api/season/{advance,status}
- Fix: BIGINT seed column (sha256-derived) clamped to 63 bits for Postgres
- 38 tests passing

+++ backend/app/models/game.py
@@ -1,4 +1,4 @@
-from sqlalchemy import ForeignKey, Integer, String
+from sqlalchemy import BigInteger, ForeignKey, Integer, String
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.db import Base
@@ -18,4 +18,4 @@ class Game(Base):
     home_shots: Mapped[int | None] = mapped_column(Integer, nullable=True)
     away_shots: Mapped[int | None] = mapped_column(Integer, nullable=True)
     result_type: Mapped[str | None] = mapped_column(String(3), nullable=True)
-    seed: Mapped[int | None] = mapped_column(Integer, nullable=True)
+    seed: Mapped[int | None] = mapped_column(BigInteger, nullable=True)
